+++ src/Donations/Endpoints/UpdateDonationAttributes/IdAttribute.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Give\Donations\Endpoints\DonationDetailsAttributes;
+namespace Give\Donations\Endpoints\UpdateDonationAttributes;
 
 use WP_Error;
